moved related components into separate folders
