+++ client/components/header/header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './styles/header.css';
+import '../styles/header.css';
 
 export default function Header(props) {
   return (
